requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
billiard==3.6.4.0
bitarray==2.7.3
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cytoolz==0.12.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
ethers==0.1.1
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-bigquery==3.5.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
jsonify==0.5
jsonschema==4.17.3
kombu==5.2.4
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.4
mysqlclient==2.1.1
netaddr==0.8.0
packaging==23.0
parsimonious==0.8.1
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.17
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
rlp==2.0.1
rsa==4.9
six==1.16.0
toolz==0.12.0
urllib3==1.26.14
varint==1.0.2
vine==5.0.0
wcwidth==0.2.6
web3==5.31.3
websocket-client==1.5.1
websockets==9.1
yarl==1.8.2
